+++ src/diffusers/schedulers/scheduling_ddpm_parallel.py
@@ -293,7 +293,7 @@ class DDPMParallelScheduler(SchedulerMixin, ConfigMixin):
                     f" {self.config.num_train_timesteps}."
                 )
 
-            timesteps = np.array(timesteps, dtype=np.int32)
+            timesteps = np.array(timesteps, dtype=np.int64)
             self.custom_timesteps = True
         else:
             if num_inference_steps > self.config.num_train_timesteps:
@@ -312,19 +312,19 @@ class DDPMParallelScheduler(SchedulerMixin, ConfigMixin):
                     np.linspace(0, self.config.num_train_timesteps - 1, num_inference_steps)
                     .round()[::-1]
                     .copy()
-                    .astype(np.int32)
+                    .astype(np.int64)
                 )
             elif self.config.timestep_spacing == "leading":
                 step_ratio = self.config.num_train_timesteps // self.num_inference_steps
                 # creates integer timesteps by multiplying by ratio
                 # casting to int to avoid issues when num_inference_step is power of 3
-                timesteps = (np.arange(0, num_inference_steps) * step_ratio).round()[::-1].copy().astype(np.int32)
+                timesteps = (np.arange(0, num_inference_steps) * step_ratio).round()[::-1].copy().astype(np.int64)
                 timesteps += self.config.steps_offset
             elif self.config.timestep_spacing == "trailing":
                 step_ratio = self.config.num_train_timesteps / self.num_inference_steps
                 # creates integer timesteps by multiplying by ratio
                 # casting to int to avoid issues when num_inference_step is power of 3
-                timesteps = np.round(np.arange(self.config.num_train_timesteps, 0, -step_ratio)).astype(np.int32)
+                timesteps = np.round(np.arange(self.config.num_train_timesteps, 0, -step_ratio)).astype(np.int64)
                 timesteps -= 1
             else:
                 raise ValueError(
